He quitado comentarios innecesarios. He borrado una funcion y su posterior readaptacion en el menu He probado a meter un metodo que detecte que solo funcionara en Windows(puede que algun dia lo haga en otros) Edicion de formato en los inputs para nombres de fichero Corrección de errores a la hora de introducir una opcion numerica

+++ CUERPO.py
@@ -6,40 +6,51 @@
 #-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 import os
 import shutil
+import platform
 from os import listdir
-#comentario random wjnmsioudfheiouhjfeiudoshniuerdhjniulfjerdhoiulhjfnerilohn
-#COMENTARIO DE PRUEBA DE PYCHARM DESDE MI UBUNTU
+
 def Menu():
     print("1.Crear carpeta de comunidad. ")
     print("2.Borrar carpeta de comunidad.")
     print("3.Consultar carpeta.")
-    print("4.Acceder a una carpeta(se mostrara una lista de los archivos).")
-    print("5.Salir")
+    print("4.Salir")
 
 def CrearCarpeta():
-    nombre=input('Introduce el nombre: ')
+    nombre=input('Introduce el nombre: ').title()
     ruta=r'C:\prueba'
     carpeta=ruta+'\\'+nombre
-    if not os.path.exists(carpeta):
-        os.makedirs(carpeta)
+    if platform.system()=="Windows":
+        if not os.path.exists(carpeta):
+            os.makedirs(carpeta)
+    else:
+        print("Error. Accion no compatible con sistema. Proximamente en actualizaciones.")
+        pass
+
 
 
 def BorrarCarpeta():
 
-    nombre=input('Nombre de la carpeta a borrar: ')
+    nombre=input('Nombre de la carpeta a borrar: ').title()
     directorio=r'C:\prueba'
     carpeta=directorio+'\\'+nombre
-    if(os.path.exists(carpeta)):
-        shutil.rmtree(carpeta)
+    if platform.system()=="Windows":
+        if (os.path.exists(carpeta)):
+            shutil.rmtree(carpeta)
+    else:
+        print("Error. Accion no compatible con sistema. Proximamente en actualizaciones.")
+        pass
+
 
 
 def AbrirCarpeta():
-    for nombre in listdir("."):
-        print(nombre)
-
-
-def Acceso():
-    ruta=input('Introduce la ruta de la carpeta: ')
+    nombre = input('Nombre de la carpeta a borrar: ').title()
+    directorio = r'C:\prueba'
+    carpeta = directorio + '\\' + nombre
+    if platform.system()=="Windows":
+        os.startfile(carpeta)
+    else:
+        print("Error. Accion no compatible con sistema. Proximamente en actualizaciones.")
+        pass
 
 def seleccion(opcion):
     if opcion==1:
@@ -52,20 +63,18 @@ def seleccion(opcion):
         #abrimos la carpeta
         AbrirCarpeta()
     elif opcion==4:
-        #Accedemos a una carpeta
-        Acceso()
-    elif opcion==5:
+        print("Saliendo del programa")
         pass
 
 
 def opcion():
-    opcion=int(input('Introduce la opcion: '))
-    while opcion>=1 and opcion<=5:
-        if(opcion.isdigit()==True):
-            seleccion(opcion)
-        else:
-            print("ERROR: Introduzca una opcion numerica.")
-            pass
-print("Bienvenido a la base de datos educacional. Seleccione una de las opciones: ")
+    opcion=eval(input('Introduce la opcion: '))
+    if opcion<1 or opcion>4:
+        print("Error: revisa rango numerico de respuestas.")
+    else:
+        seleccion(opcion)
+
+
+print("Acceso a Manejo de Ficheros del Centro Educacional. Seleccione una de las opciones: ")
 Menu()
 opcion()
